feat: add initialPreferences
