+++ lib/user-preference-store/utils.ts
@@ -1,0 +1,47 @@
+export type LayoutType = "sidebar" | "fullscreen"
+
+export type UserPreferences = {
+  layout: LayoutType
+  promptSuggestions: boolean
+  showToolInvocations: boolean
+  showConversationPreviews: boolean
+  multiModelEnabled: boolean
+  hiddenModels: string[]
+}
+
+export const defaultPreferences: UserPreferences = {
+  layout: "fullscreen",
+  promptSuggestions: true,
+  showToolInvocations: true,
+  showConversationPreviews: true,
+  multiModelEnabled: false,
+  hiddenModels: [],
+}
+
+// Helper functions to convert between API format (snake_case) and frontend format (camelCase)
+export function convertFromApiFormat(apiData: any): UserPreferences {
+  return {
+    layout: apiData.layout || "fullscreen",
+    promptSuggestions: apiData.prompt_suggestions ?? true,
+    showToolInvocations: apiData.show_tool_invocations ?? true,
+    showConversationPreviews: apiData.show_conversation_previews ?? true,
+    multiModelEnabled: apiData.multi_model_enabled ?? false,
+    hiddenModels: apiData.hidden_models || [],
+  }
+}
+
+export function convertToApiFormat(preferences: Partial<UserPreferences>) {
+  const apiData: any = {}
+  if (preferences.layout !== undefined) apiData.layout = preferences.layout
+  if (preferences.promptSuggestions !== undefined)
+    apiData.prompt_suggestions = preferences.promptSuggestions
+  if (preferences.showToolInvocations !== undefined)
+    apiData.show_tool_invocations = preferences.showToolInvocations
+  if (preferences.showConversationPreviews !== undefined)
+    apiData.show_conversation_previews = preferences.showConversationPreviews
+  if (preferences.multiModelEnabled !== undefined)
+    apiData.multi_model_enabled = preferences.multiModelEnabled
+  if (preferences.hiddenModels !== undefined)
+    apiData.hidden_models = preferences.hiddenModels
+  return apiData
+}
